support save awq
